requirements.txt
fastapi[standard]==0.115.12
sqlmodel==0.0.24
passlib==1.7.4
pyjwt==2.10.1
httpx==0.28.1
pytest==8.3.5
cryptography==44.0.2